+++ lib/install.ps1
@@ -737,16 +737,28 @@ function run_installer($fname, $manifest, $architecture, $dir, $global) {
     # MSI or other installer
     $msi = msi $manifest $architecture
     $installer = installer $manifest $architecture
-    if ($installer.script) {
-        Write-UserMessage -Message 'Running installer script...' -Output:$false
-        Invoke-Expression (@($installer.script) -join "`r`n")
-        return
+    $script = $installer.script
+
+    if ($script) {
+        # Skip installer if installer property specifies only 1 property == script
+        # If there is file or args next to the script, both should be called, script first
+        $props = @($installer.PSObject.Properties | Where-Object -Property 'MemberType' -EQ -Value 'NoteProperty' | Select-Object -ExpandProperty 'Name')
+        if ($props -and ($props.Count -eq 1) -and ($props -contains 'script')) {
+            $skipInstaller = $true
+        }
     }
 
     if ($msi) {
         install_msi $fname $dir $msi
-    } elseif ($installer) {
+    } elseif ($installer -and !$skipInstaller) {
         install_prog $fname $dir $installer $global
+    }
+
+    # Run installer.script after installer.file
+    # This allow to modify files after installer was executed and before binaries/shortcuts are created/linked
+    if ($script) {
+        Write-UserMessage -Message 'Running installer script...' -Output:$false
+        Invoke-Expression (@($script) -join "`r`n")
     }
 }
 
@@ -797,17 +809,20 @@ function install_prog($fname, $dir, $installer, $global) {
     $arg = @(args $installer.args $dir $global)
 
     if ($prog.EndsWith('.ps1')) {
+        Write-UserMessage -Message "Running installer file '$prog'" -Output:$false
         & $prog @arg
-        # TODO: Handle $LASTEXITCODE
+        if ($LASTEXITCODE -ne 0) {
+            throw [ScoopException] "Installation failed with exit code $LASTEXITCODE" # TerminatingError thrown
+        }
     } else {
         $installed = Invoke-ExternalCommand $prog $arg -Activity 'Running installer...'
         if (!$installed) {
             throw [ScoopException] "Installation aborted. You might need to run 'scoop uninstall $app' before trying again." # TerminatingError thrown
         }
-
-        # Don't remove installer if "keep" flag is set to true
-        if ($installer.keep -ne 'true') { Remove-Item $prog }
     }
+
+    # Do not remove installer if "keep" flag is set to true
+    if ($installer.keep -ne 'true') { Remove-Item $prog }
 }
 
 function run_uninstaller($manifest, $architecture, $dir) {